requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-environ==0.4.5
future==0.18.2
idna==2.10
line-bot-sdk==1.18.0
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2

uwsgi==2.0.18
